+++ tests/testthat/test_theme_ggstatsplot.R
@@ -50,6 +50,7 @@ testthat::test_that(
 testthat::test_that(
   desc = "`theme_pie()` works",
   code = {
+    testthat::skip_on_cran()
     set.seed(123)
     library(ggplot2)